refactored extraction methods, poly toe and lcp toe not working yet
Extraction functions now assume the data they are given (xy_data, crest data, etc) has the desired index assigned already. This made it possible to reduce some repeated code in each function that would assign the index. Also, the heel and IP functions  have a more readable way of finding the minimum x / y value in each group after filtering, but it seems this method may be slower so it may be reverted.