Swap axes of xy_data to make rolling window possible, this causes broadcasting problem
Each variable (x, y, rr) columns is now assigned a sequence of subcolumns (x1, x2, x3, etc) so that rolling calculations can be done along axis 0 rather than axis 1 (rolling along axis 1 in pandas is less tested and apparently not working properly in all cases). However, when a filtering operation such as xy_data[xy_data["x"] > 5] is done, the dimensions of the boolean mask are only a factor of the whole dataframe (i.e. one third of the columns, since we're masking 1 out of the 3 variables). It should be possible to broadcast this across the DataFrame, but I'm not sure how to do it.